requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.6
django-crispy-forms==1.12.0
django-extensions==3.1.3
djangorestframework==3.12.4
dnspython==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
PyJWT==1.7.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.3.0
